make it possible to specify 'or' on a requirement, add button to duplicate requirement

+++ lib/src/noita_random.cpp
@@ -165,7 +165,7 @@ extern "C"
         return Random(a, b);
     }
 
-    double ProceduralRandomi(double x, double y, double a, double b) {
+    int ProceduralRandomi(double x, double y, double a, double b) {
         SetRandomSeed(x, y);
         return (int)Random(a, b);
     }
